backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pandas==2.1.3
PyPDF2==3.0.1
python-docx==1.1.0
Pillow==10.1.0
pytesseract==0.3.10
spacy==3.7.2
openai==1.3.7
google-generativeai==0.3.2
scikit-learn==1.3.2
azure-cognitiveservices-vision-computervision==0.9.0
azure-storage-blob==12.19.0
python-dotenv==1.0.0
aiofiles==23.2.1
pydantic==2.5.0
httpx==0.25.2
numpy==1.25.2
regex==2023.10.3
